Design: 디자이너 게시물 디테일 뷰 디자인 수정
